refactor: move format_table function to utils
