+++ tutor/utils.py
@@ -345,3 +345,31 @@ def is_http(url: str) -> bool:
     Basic test to check whether a string is a web URL. Use only for basic use cases.
     """
     return re.match(r"^https?://", url) is not None
+
+
+def format_table(rows: List[Tuple[str, ...]], separator: str = "\t") -> str:
+    """
+    Format a list of values as a tab-separated table. Column sizes are determined such
+    that row values are vertically aligned.
+    """
+    formatted = ""
+    if not rows:
+        return formatted
+    columns_count = len(rows[0])
+    # Determine each column size
+    col_sizes = [1] * columns_count
+    for row in rows:
+        for c, value in enumerate(row):
+            col_sizes[c] = max(col_sizes[c], len(value))
+    # Print all values
+    for r, row in enumerate(rows):
+        for c, value in enumerate(row):
+            if c < len(col_sizes) - 1:
+                formatted += f"{value:{col_sizes[c]}}{separator}"
+            else:
+                # The last column is not left-justified
+                formatted += f"{value}"
+        if r < len(rows) - 1:
+            # Append EOL at all lines but the last one
+            formatted += "\n"
+    return formatted
